Refactoring : Change upnp entity package
