+++ jpsonic-main/src/main/java/com/tesshu/jpsonic/service/upnp/processor/ArtistByFolderProc.java
@@ -26,8 +26,8 @@ import com.tesshu.jpsonic.dao.AlbumDao;
 import com.tesshu.jpsonic.dao.ArtistDao;
 import com.tesshu.jpsonic.domain.Artist;
 import com.tesshu.jpsonic.service.upnp.ProcId;
-import com.tesshu.jpsonic.service.upnp.composite.ArtistOrAlbum;
-import com.tesshu.jpsonic.service.upnp.composite.FolderOrArtist;
+import com.tesshu.jpsonic.service.upnp.processor.composite.ArtistOrAlbum;
+import com.tesshu.jpsonic.service.upnp.processor.composite.FolderOrArtist;
 import org.jupnp.support.model.DIDLContent;
 import org.jupnp.support.model.container.Container;
 import org.springframework.stereotype.Service;
