+++ jpsonic-main/src/main/java/com/tesshu/jpsonic/service/upnp/processor/RandomSongByFolderArtistProc.java
@@ -26,8 +26,8 @@ import com.tesshu.jpsonic.dao.ArtistDao;
 import com.tesshu.jpsonic.service.SearchService;
 import com.tesshu.jpsonic.service.SettingsService;
 import com.tesshu.jpsonic.service.upnp.ProcId;
-import com.tesshu.jpsonic.service.upnp.composite.ArtistOrSong;
-import com.tesshu.jpsonic.service.upnp.composite.FolderOrArtist;
+import com.tesshu.jpsonic.service.upnp.processor.composite.ArtistOrSong;
+import com.tesshu.jpsonic.service.upnp.processor.composite.FolderOrArtist;
 import org.jupnp.support.model.DIDLContent;
 import org.jupnp.support.model.container.Container;
 import org.springframework.stereotype.Service;
