+++ jpsonic-main/src/test/java/com/tesshu/jpsonic/service/upnp/processor/FolderOrArtistLogicTest.java
@@ -39,7 +39,7 @@ import com.tesshu.jpsonic.service.PlayerService;
 import com.tesshu.jpsonic.service.SettingsService;
 import com.tesshu.jpsonic.service.TranscodingService;
 import com.tesshu.jpsonic.service.upnp.ProcId;
-import com.tesshu.jpsonic.service.upnp.composite.FolderOrArtist;
+import com.tesshu.jpsonic.service.upnp.processor.composite.FolderOrArtist;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.jupnp.support.model.container.Container;
