+++ jpsonic-main/src/test/java/com/tesshu/jpsonic/service/upnp/processor/RandomSongByFolderArtistProcTest.java
@@ -43,8 +43,8 @@ import com.tesshu.jpsonic.service.PlayerService;
 import com.tesshu.jpsonic.service.SearchService;
 import com.tesshu.jpsonic.service.SettingsService;
 import com.tesshu.jpsonic.service.TranscodingService;
-import com.tesshu.jpsonic.service.upnp.composite.ArtistOrSong;
-import com.tesshu.jpsonic.service.upnp.composite.FolderOrArtist;
+import com.tesshu.jpsonic.service.upnp.processor.composite.ArtistOrSong;
+import com.tesshu.jpsonic.service.upnp.processor.composite.FolderOrArtist;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
